% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAvg.R
\name{findAvg}
\alias{findAvg}
\title{Uses inported Rs data to find the average of each year}
\usage{
findAvg(data, idx)
}
\arguments{
\item{data}{list from imported matlab data}

\item{idx}{the index of the year you like
1 = 2014
2 = 2015
3 = 2016
4 = 2017
5 = 2018}
}
\value{
average data
}
\description{
Uses inported Rs data to find the average of each year
}
\examples{
findAvg(rsIn)
}
