% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRs.R
\docType{package}
\name{plotRs}
\alias{plotRs}
\alias{plotRs-package}
\title{plotRs: a package for plotting soil respiration}
\description{
Extracts data from matlab and then finds the daily average
}
\references{
\url{https://github.com/may118/My_First_Repository}
}
\author{
David Ma, School of Geography and Earth Sciences, McMaster University \email{may118@mcmaster.ca}
}
