% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Uses R.matlab to read .mat data file into R}
\usage{
readData(name)
}
\arguments{
\item{name}{the name of a matlab file}
}
\value{
The R output version of data
}
\description{
Uses R.matlab to read .mat data file into R
}
\examples{
data = readData('inputData.mat')
}
